phish-sim/visual/requirements.txt
# SPDX-License-Identifier: MIT
# Copyright (c) 2024 Phish-Sim Project

# Web automation and screenshot capture
playwright==1.40.0
selenium==4.15.2
beautifulsoup4==4.12.2
lxml==4.9.3

# Computer vision and image processing
opencv-python==4.8.1.78
Pillow==10.1.0
scikit-image==0.22.0
imageio==2.31.6

# Deep learning for visual analysis
torch==2.1.0
torchvision==0.16.0
tensorflow==2.15.0
keras==2.15.0

# Image analysis and feature extraction
numpy==1.24.3
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0

# DOM analysis and parsing
html5lib==1.1
cssselect==1.2.0
requests==2.31.0
urllib3==2.1.0

# Template matching and similarity
scikit-learn==1.3.2
opencv-contrib-python==4.8.1.78

# API and web framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
click==8.1.7
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0